Remove old TextureSampler
