+++ Editor/Foundation/NativeTextureSampler.cs.meta
